+++ .storybook/main.ts
@@ -1,5 +1,6 @@
 import type { StorybookConfig } from '@storybook/react-webpack5';
 import path from 'path';
+import React from 'react';
 
 const config: StorybookConfig = {
     stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
@@ -15,15 +16,17 @@ const config: StorybookConfig = {
         name: '@storybook/react-webpack5',
         options: {}
     },
-    staticDirs: ['../public'],
-    webpackFinal: async (storybookConfig) => {
-        // 이름 변경
-        storybookConfig.resolve = storybookConfig.resolve || {};
-        storybookConfig.resolve.alias = {
-            ...storybookConfig.resolve.alias,
-            '@': path.resolve(__dirname, '../src')
+    staticDirs: ['../public'], // 정적 파일 디렉토리
+    webpackFinal: async (config) => {
+        // Webpack alias 설정 추가
+        config.resolve = config.resolve || {};
+        config.resolve.alias = {
+            ...config.resolve.alias,
+            '@': path.resolve(__dirname, '../src'), // '@'를 src 폴더로 매핑
+            '@/components': path.resolve(__dirname, '../src/components') // '@/components'를 src/components로 매핑
         };
-        return storybookConfig;
+
+        return config;
     }
 };
