+++ .storybook/preview.tsx
@@ -1,0 +1,25 @@
+import type { Preview } from '@storybook/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import React from 'react';
+
+const queryClient = new QueryClient();
+
+const preview: Preview = {
+    parameters: {
+        controls: {
+            matchers: {
+                color: /(background|color)$/i,
+                date: /Date$/i
+            }
+        }
+    },
+    decorators: [
+        (Story) => (
+            <QueryClientProvider client={queryClient}>
+                <Story />
+            </QueryClientProvider>
+        )
+    ]
+};
+
+export default preview;
